fluids: Move diff filter enums to separate header file
- So that other QFunctions can use them for indexing into differential
  filtered fields
